Bugfix: prevent name from being set to custom
This is necessitated by what appears to be a bug in dm-core.  Apparently
a custom accessor which will variably return the value of one of two
columns can actually set the value of a column due to the way
dm-core/property_set works.  For a concrete example, see the
implementation prior to this commit of the Session#name method.

- provide a generic accessor method that returns either custom (if
  non-nil) or name as a last resort
- move to using this display_name method
- refactored code and specs to suit this new approach